Move all i18n related stuff from Modal to messageUtils
